# backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EasyClaim API",
    "version": "2.0.0-roles",
    "description": "Generated from the mounted routes of backend/src/index.ts (see docs/API_CONTRACT.md). Roles: CUSTOMER, INSURER_ADMIN, SUPERADMIN. Every route needs a Bearer JWT except /auth/login and /auth/register. Served at /openapi.json and /swagger only when ENVIRONMENT is development or demo."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Decide": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "outcome": {
            "type": "string",
            "enum": [
              "Approved",
              "Rejected"
            ]
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "approvedAmountCents": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000000,
            "description": "Approved only; omitted = full claimed amount; never above the claimed amount"
          }
        },
        "required": [
          "outcome",
          "reason"
        ]
      },
      "Transition": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "transitioned"
            ]
          },
          "claimId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,64}$"
          },
          "from": {
            "type": "string",
            "enum": [
              "Draft",
              "Submitted",
              "Verified",
              "Screening",
              "Review",
              "Decision",
              "Paid",
              "Info Needed",
              "Appeal",
              "Withdrawn",
              "Expired"
            ]
          },
          "to": {
            "type": "string",
            "enum": [
              "Draft",
              "Submitted",
              "Verified",
              "Screening",
              "Review",
              "Decision",
              "Paid",
              "Info Needed",
              "Appeal",
              "Withdrawn",
              "Expired"
            ]
          }
        }
      },
      "RiskSignals": {
        "type": "object",
        "nullable": true,
        "description": "Advisory screening context produced offline (quantum/). Never changes stage, status or payout; no route accepts it.",
        "properties": {
          "classicalAnomaly": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "quantumAnomaly": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "interpretation": {
            "type": "string",
            "enum": [
              "NORMAL",
              "UNUSUAL",
              "HIGH_ANOMALY"
            ]
          },
          "anomalyBand": {
            "type": "string",
            "enum": [
              "NORMAL",
              "ELEVATED",
              "HIGH"
            ]
          },
          "screeningRecommendation": {
            "type": "string",
            "enum": [
              "STANDARD_REVIEW",
              "REVIEW_REQUIRED"
            ]
          },
          "explanation": {
            "type": "string"
          },
          "versions": {
            "type": "object",
            "properties": {
              "model": {
                "type": "string"
              },
              "features": {
                "type": "string"
              },
              "kernel": {
                "type": "string"
              },
              "screening": {
                "type": "string"
              }
            }
          },
          "signalDigest": {
            "type": "string"
          },
          "modelVersion": {
            "type": "string"
          },
          "execution": {
            "type": "string",
            "enum": [
              "simulator",
              "hardware"
            ]
          },
          "computedAt": {
            "type": "string"
          },
          "advisory": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/claims/{claimId}/verify": {
      "post": {
        "tags": [
          "Insurer"
        ],
        "summary": "Submitted → Verified (INSURER_ADMIN of the tenant)",
        "parameters": [
          {
            "name": "claimId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transition"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "not_found (also for claims the caller may not see)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "illegal_transition / stale_state / claim_not_editable / already_paid / decision_integrity_failed / ...",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/claims/{claimId}/screen": {
      "post": {
        "tags": [
          "Insurer"
        ],
        "summary": "Verified → Screening; returns the advisory risk signal (no request body is read)",
        "parameters": [
          {
            "name": "claimId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Transition"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "riskSignals": {
                          "$ref": "#/components/schemas/RiskSignals"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "not_found (also for claims the caller may not see)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "illegal_transition / stale_state / claim_not_editable / already_paid / decision_integrity_failed / ...",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/claims/{claimId}/review": {
      "post": {
        "tags": [
          "Insurer"
        ],
        "summary": "Screening → Review, Appeal → Review (INSURER_ADMIN)",
        "parameters": [
          {
            "name": "claimId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transition"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "not_found (also for claims the caller may not see)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "illegal_transition / stale_state / claim_not_editable / already_paid / decision_integrity_failed / ...",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/claims/{claimId}/request-info": {
      "post": {
        "tags": [
          "Insurer"
        ],
        "summary": "Screening / Review → Info Needed",
        "parameters": [
          {
            "name": "claimId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transition"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "not_found (also for claims the caller may not see)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "illegal_transition / stale_state / claim_not_editable / already_paid / decision_integrity_failed / ...",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/claims/{claimId}/decide": {
      "post": {
        "tags": [
          "Decisions"
        ],
        "summary": "Review → Decision (INSURER_ADMIN); ML-DSA-65 signed decision record",
        "parameters": [
          {
            "name": "claimId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,64}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Decide"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "validation_failed / client_supplied_fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "not_found (also for claims the caller may not see)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "illegal_transition / stale_state / claim_not_editable / already_paid / decision_integrity_failed / ...",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "business rule (see API_CONTRACT.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "storage_unavailable / integrity_unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/claims/{claimId}/pay": {
      "post": {
        "tags": [
          "Payouts"
        ],
        "summary": "Decision(Approved) → Paid, SIMULATED (INSURER_ADMIN); empty body; verifies the decision signature and destination",
        "parameters": [
          {
            "name": "claimId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,64}$"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "validation_failed / client_supplied_fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "not_found (also for claims the caller may not see)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "illegal_transition / stale_state / claim_not_editable / already_paid / decision_integrity_failed / ...",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}